+++ spec/kitchen_spec/directions/move_solutions_to_answer_key/move_solutions_from_exercise_section_spec.rb
@@ -65,7 +65,7 @@ RSpec.describe Kitchen::Directions::MoveSolutionsFromExerciseSection do
     end
 
     it 'bakes' do
-      described_class.v1(chapter: section_with_exercises, append_to: append_element, section_class: 'exercise-section', in_appendix: true)
+      described_class.v1(chapter: section_with_exercises, append_to: append_element, section_class: 'exercise-section', options: { in_appendix: true })
       expect(append_element).to match_snapshot_auto
     end
   end
@@ -77,6 +77,11 @@ RSpec.describe Kitchen::Directions::MoveSolutionsFromExerciseSection do
 
   it 'bakes a numbered section' do
     described_class.v1(chapter: section_with_exercises, append_to: append_element, section_class: 'another-exercise-section', title_number: '3.4')
+    expect(append_element).to match_snapshot_auto
+  end
+
+  it 'bakes without section title' do
+    described_class.v1(chapter: section_with_exercises, append_to: append_element, section_class: 'exercise-section', options: { add_title: false })
     expect(append_element).to match_snapshot_auto
   end
 end